Members Pages Modified
Members now shows a tabled list of all the members in the GymSane system. Each name is a hyperlink, which will allow the user to click on any given name and be able to view the details and sessions that a user is signed up for. There is also an Edit button, that will allow the user to change any details regarding the member.