Sessions Pages Modified
Sessions now shows a tabled list of all the sessions in the GymSane system. Each name is a hyperlink, which will allow the user to click on any given name and be able to view the details and members that are signed up. There is also an Edit button, that will allow the user to change any details regarding the session.